+++ src/core/dfs.cc
@@ -3,7 +3,7 @@
 //
 
 #include <visualizer/sketchpad.h>
-#include "core/dfs.h"
+#include <core/dfs.h>
 
 namespace graph_algorithm {
 
@@ -14,7 +14,7 @@ DFS::DFS(Graph *board_graph) {
   found_destination_ = false;
 }
 
-void DFS::RecursiveDFS(int node, size_t end_row, size_t end_col,
+void DFS::RecursiveDFS(size_t node, size_t end_row, size_t end_col,
                        vector<bool> &visited_nodes) {
   if (!found_destination_) {
     visited_nodes[node] = true;
@@ -26,7 +26,7 @@ void DFS::RecursiveDFS(int node, size_t end_row, size_t end_col,
 
     } else {
       // recursively iterate through the current node's adjacent nodes
-      for (int i : board_graph_->GetNodes().at(node)->adjacent) {
+      for (size_t i : board_graph_->GetNodes().at(node)->adjacent) {
         if (!visited_nodes[i]) {
           RecursiveDFS(i, end_row, end_col, visited_nodes);
         }
@@ -47,7 +47,7 @@ vector<vector<int>> DFS::RunDFS(size_t end_row, size_t end_col) {
 
   // update the output board to show the nodes we've visited from the start to
   // find the end destination
-  for (int i = 0; i < dimension * dimension; i++) {
+  for (size_t i = 0; i < dimension * dimension; i++) {
     if (visited_nodes[i]) {
       Graph::Node *node = board_graph_->GetNodes().at(i);
       output_board[node->row][node->col] = graph_algorithm::kPath;
